Move check read only to helper
