+++ open_alchemy/helpers/__init__.py
@@ -1,6 +1,7 @@
 """Helper functions."""
 # pylint: disable=useless-import-alias
 
+from .check_read_only import check_read_only as check_read_only
 from .define_all import define_all as define_all
 from .get_ext_prop import get_ext_prop as get_ext_prop
 from .merge_all_of import merge_all_of as merge_all_of
